Update README file
The instructions at the top were a bit confusing. Tried to clarify how to 1. sign up or sign I  2. install stepzen, and 3. stepzen import ...

Sending users to the hello world to get set up is confusing. I know `stepzen import` is not part of the "Get Set Up " docs section yet, but will need to be to go along with this release (we have a docs ticket for it)